Add files via upload
Project Overview
In this project, you will analyze a real dataset about current Udacity students across a number of programs, so it isn't perfect. It is a little messy (some things are input incorrectly, others are missing). You will need to decide how to analyze the data and then communicate your findings about it. You will use spreadsheets to make your analysis easier (please do not analyze by hand - it will take forever!).

Prepare with the descriptive statistics and spreadsheets lessons leading up to the project.

What do I need to install?
You may use any spreadsheet application you like. This includes Google Sheets, Microsoft Excel, etc.

Why this Project?
This project will introduce you to the data analysis process that you will be using throughout the rest of the Nanodegree program. In this project, you will go through the process of calculating summary statistics and building visual methods of spreadsheet data to find interesting information out about Udacity students - students just like you. The goal is that you not only can perform an analysis, but also that you can communicate the results in a way that is clear and compelling.

What skills will I use?
After completing the lessons and the project, you should feel comfortable in your ability to:

Identify different datatypes to determine what analyses are possible.
Calculate measures of center and spread to describe quantitative data.
Use a variety of spreadsheet functionality to gain insights from data.
Build charts and graphs to communicate the results of your analysis
Be comfortable posing questions that can be answered with a given dataset and then answering those questions
Analyze a future dataset using Excel
Project Details
The data used for this project was collected from Udacity students who responded to a survey. This project is made of two parts. For both parts you will be using the same dataset, which you can find in the Supporting Materials as Surveydata3.csv at the bottom of this page or as a google sheet here. You will notice that some information has been removed to assure the privacy of the student respondents.

The first part of the project is a set of quiz questions, which you will find in the upcoming concepts. These concepts are aimed to help you get familiar with the dataset and test that you have mastered the core concepts in the previous lessons. Correctly answering each of the quiz questions will assure you are on the right track before you dive into the second part of the project. This part of the project will not be submitted for review.


The second part of your project is the portion you will turn in for review. You will need to create a presentation to be reviewed. The details of this submission are detailed in the last page in this lesson. Pay attention to the details of the Rubric to assure you have all deliverables. In order to review your presentation, you will need to save your slides as a PDF.